Refactoring Task to sep. file
